// app/agent/(AgentLayout)/transactions/types.ts
export interface Transaction {
  id: string;
  transaction_date?: string;
  transaction_stage?: string;
  transaction_status?: string;
  transaction_id?: string;
  reference_number?: string;
  transaction_type?: string;
  group?: string;
  referenceNumber?: string;
  type?: string;
  stage?: string;
  status?: string;
  date?: string;
}